Refactor to com.ywesee.java.yopenedi
