+++ java/cli/src/test/java/com/ywesee/java/yopenedi/cli/AppTest.java
@@ -1,0 +1,13 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package com.ywesee.java.yopenedi.cli;
+
+import org.junit.Test;
+
+public class AppTest {
+    @Test public void testAppHasAGreeting() {
+        App classUnderTest = new App();
+        //assertNotNull("app should have a greeting", classUnderTest.getGreeting());
+    }
+}
