+++ java/converter/src/main/java/com/ywesee/java/yopenedi/converter/ContactDetail.java
@@ -1,0 +1,8 @@
+package com.ywesee.java.yopenedi.converter;
+
+public class ContactDetail {
+    public String name;
+    public String phone;
+    public String email;
+    public String fax;
+}
