+++ java/converter/src/main/java/com/ywesee/java/yopenedi/converter/OrderItem.java
@@ -1,0 +1,12 @@
+package com.ywesee.java.yopenedi.converter;
+
+import java.math.BigDecimal;
+
+public class OrderItem {
+    public String ean;
+    public String descriptionShort;
+    public String descriptionLong;
+    public BigDecimal quantity;
+    public String quantityUnit;
+    public BigDecimal price;
+}
