+++ java/converter/src/test/java/com/ywesee/java/yopenedi/converter/LibraryTest.java
@@ -1,0 +1,13 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package com.ywesee.java.yopenedi.converter;
+
+import org.junit.Test;
+import static org.junit.Assert.*;
+
+public class LibraryTest {
+    @Test public void testSomeLibraryMethod() {
+        assertTrue("someLibraryMethod should return 'true'", true);
+    }
+}
