--- java/openedi-converter/src/main/java/openedi/converter/ContactDetail.java
@@ -1,8 +1,0 @@
-package openedi.converter;
-
-public class ContactDetail {
-    public String name;
-    public String phone;
-    public String email;
-    public String fax;
-}
