--- java/openedi-converter/src/main/java/openedi/converter/OrderItem.java
@@ -1,12 +1,0 @@
-package openedi.converter;
-
-import java.math.BigDecimal;
-
-public class OrderItem {
-    public String ean;
-    public String descriptionShort;
-    public String descriptionLong;
-    public BigDecimal quantity;
-    public String quantityUnit;
-    public BigDecimal price;
-}
