--- java/openedi-converter/src/test/java/openedi/converter/LibraryTest.java
@@ -1,13 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package openedi.converter;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-public class LibraryTest {
-    @Test public void testSomeLibraryMethod() {
-        assertTrue("someLibraryMethod should return 'true'", true);
-    }
-}
