--- java/openedi-exe/build.gradle
@@ -1,42 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Java project to get you started.
- * For more details take a look at the Java Quickstart chapter in the Gradle
- * User Manual available at https://docs.gradle.org/6.0.1/userguide/tutorial_java_projects.html
- */
-
-plugins {
-    // Apply the java plugin to add support for Java
-    id 'java'
-
-    // Apply the application plugin to add support for building a CLI application.
-    id 'application'
-    id 'com.github.johnrengelman.shadow' version '5.2.0'
-}
-
-repositories {
-    // Use jcenter for resolving dependencies.
-    // You can declare any Maven/Ivy/file repository here.
-    jcenter()
-}
-
-dependencies {
-    // This dependency is used by the application.
-    implementation 'com.google.guava:guava:28.0-jre'
-    compile project(':openedi-converter')
-
-    implementation 'commons-cli:commons-cli:1.4'
-
-    // Use JUnit test framework
-    testImplementation 'junit:junit:4.12'
-}
-
-application {
-    // Define the main class for the application.
-    mainClassName = 'openedi.exe.App'
-}
-
-run{
-    standardInput = System.in
-}
